Module L2__Hypothesis

Modules for constructing hypotheses.

module StaticDistance : sig ... end

Static distance coordinates.

module Symbol : sig ... end

Symbols are constant strings with a fast comparison function. Used as identifiers.

module Hole : sig ... end

Holes are subproblems within a Hypothesis. They have a type, a type context, and a symbol which defines the set of expressions which can be used to fill the hole.

module Skeleton : sig ... end

Skeletons are expressions with Holes that are annotated with Specifications.

module CostModel : sig ... end

A CostModel assigns a cost to each variant of a skeleton. The total cost is the sum of the costs of the nodes.

module PerFunctionCostModel : sig ... end

module Specification : sig ... end

module Examples : sig ... end

module FunctionExamples : sig ... end

module Inputs : sig ... end

module Hypothesis : sig ... end

Hypotheses are a further refinement of Skeletons which add additional information: cost, abstract/concrete state, and a list of holes.