Module L2__Hypothesis.StaticDistance

Static distance coordinates.

type t

include Core.Sexpable.S with type StaticDistance.t := t

val t_of_sexp : Base__.Sexp.t -> t

val sexp_of_t : t -> Base__.Sexp.t

include Core.Comparator.S with type StaticDistance.t := t

type comparator_witness

val comparator : (t, comparator_witness) Core.Comparator.comparator

val map_to_string : ('a -> string) -> (t, 'a, _) Core.Map.t -> string

val increment_scope : t -> t

Increment the scope of a coordinate.

val map_increment_scope : (t, 'a, comparator_witness) Core.Map.t -> (t, 'a, comparator_witness) Core.Map.t

Increment the scope of every coordinate in a Map.t.

val create : distance:int -> index:int -> t

Create a static distance coordinate.

parameter distance

The distance of the coordinate from its binding site, in number of bindings.

parameter index

The binding index.

val distance : t -> int

Get the distance of a static distance coordinate.

val index : t -> int

Get the index of a static distance coordinate.

val args : int -> t list

Generate an arguments list with n arguments.

val to_string : t -> string

Convert static distance coordinate to string.