Module L2__Mutctx

type 'a t = 'a Core.Map.M(L2.Name).t Core.ref

val compare : a. ('a -> 'a -> int) -> 'a t -> 'a t -> Core_kernel__.Import.int

val t_of_sexp : a. (Ppx_sexp_conv_lib.Sexp.t -> 'a) -> Ppx_sexp_conv_lib.Sexp.t -> 'a t

val sexp_of_t : a. ('a -> Ppx_sexp_conv_lib.Sexp.t) -> 'a t -> Ppx_sexp_conv_lib.Sexp.t

exception UnboundError of L2.Name.t

val empty : unit -> (L2.Name.t, 'a, L2.Name.comparator_witness) Core.Map.t Core.ref

Return an empty context.

val lookup : ('a, 'b, 'c) Core.Map.t Core.ref -> 'a -> 'b Core_kernel__.Import.option

Look up an id in a context.

val lookup_exn : (L2.Name.t, 'a, 'b) Core.Map.t Core.ref -> L2.Name.t -> 'a

val bind : ('a, 'b, 'c) Core.Map.t Core.ref -> 'a -> 'b -> ('a, 'b, 'c) Core.Map.t Core.ref

Bind a type or value to an id, returning a new context.

val bind_alist : ('a, 'b, 'c) Core.Map.t Core.ref -> ('a * 'b) list -> ('a, 'b, 'c) Core.Map.t Core.ref

val unbind : ('a, 'b, 'c) Core.Map.t Core.ref -> 'a -> ('a, 'b, 'c) Core.Map.t Core.ref

Remove a binding from a context, returning a new context.

val update : ('a, 'b, 'c) Core.Map.t Core.ref -> 'a -> 'b -> unit

Bind a type or value to an id, updating the context in place.

val remove : ('a, 'b, 'c) Core.Map.t Core.ref -> 'a -> unit

Remove a binding from a context, updating the context in place.

val merge : ('a, 'b, 'c) Core.Map.t Core.ref -> ('a, 'd, 'c) Core.Map.t Core.ref -> f:(key:'a -> [ `Both of 'b * 'd | `Left of 'b | `Right of 'd ] -> 'e Core_kernel__.Import.option) -> ('a, 'e, 'c) Core.Map.t Core.ref

val merge_right : 'a t -> 'a t -> 'a t

val map : ('a, 'b, 'c) Core.Map.t Core.ref -> f:('b -> 'd) -> ('a, 'd, 'c) Core.Map.t Core.ref

val mapi : ('a, 'b, 'c) Core.Map.t Core.ref -> f:(key:'a -> data:'b -> 'd) -> ('a, 'd, 'c) Core.Map.t Core.ref

val filter : ('a, 'b, 'c) Core.Map.t Core.ref -> f:(key:'a -> data:'b -> Core_kernel__.Import.bool) -> ('a, 'b, 'c) Core.Map.t Core.ref

val filter_mapi : ('a, 'b, 'c) Core.Map.t Core.ref -> f:(key:'a -> data:'b -> 'd Core_kernel__.Import.option) -> ('a, 'd, 'c) Core.Map.t Core.ref

val equal : ('a -> 'a -> Core_kernel__.Import.bool) -> ('b, 'a, 'c) Core.Map.t Core.ref -> ('b, 'a, 'c) Core.Map.t Core.ref -> Core_kernel__.Import.bool

val keys : ('a, 'b, 'c) Core.Map.t Core.ref -> 'a Core_kernel__.Import.list

val data : ('a, 'b, 'c) Core.Map.t Core.ref -> 'b Core_kernel__.Import.list

val of_alist : ('a, 'b) Core.Map.comparator -> (('a * 'c) Core_kernel__.Import.list -> [ `Duplicate_key of 'a | `Ok of ('a, 'c, 'b) Core.Map.t ]) Core.ref

val of_alist_exn : (L2.Name.t * 'a) Core_kernel__.Import.list -> (L2.Name.t, 'a, L2.Name.comparator_witness) Core.Map.t Core.ref

val of_alist_mult : (L2.Name.t * 'a) Core_kernel__.Import.list -> (L2.Name.t, 'a Core_kernel__.Import.list, L2.Name.comparator_witness) Core.Map.t Core.ref

val of_string_map : 'a -> 'a Core.ref

val to_string_map : 'a Core.ref -> 'a

val to_alist : ('a, 'b, 'c) Core.Map.t Core.ref -> ('a * 'b) Core_kernel__.Import.list

val to_string : 'a t -> ('a -> string) -> string

val of_ctx : 'a -> 'a Core.ref

val to_ctx : 'a Core.ref -> 'a lazy_t